Extract active recording list component
